seeding/requirements.txt
PyQt5>=5.15.0
qt-material>=2.14
numpy>=1.23
opencv-python>=4.7.0
Pillow>=9.4.0
PyMuPDF>=1.22.0       # для работы с PDF (fitz)
ultralytics>=8.0.0    # YOLOv8 (если используешь детектор)
reportlab>=3.6.0      # генерация PDF-отчётов
